refactor: move combinatorics functions to particle
