+++ tests/channels/test_jpsi_to_gamma_pi0_pi0.py
@@ -6,11 +6,11 @@ import logging
 
 import pytest
 
+from expertsystem.state.particle import _create_edge_id_particle_mapping
 from expertsystem.ui import (
     InteractionTypes,
     StateTransitionManager,
 )
-from expertsystem.ui._system_control import _create_edge_id_particle_mapping
 
 
 logging.basicConfig(level=logging.ERROR)
